Fix testcases

+++ hypernets/tests/tabular/sklearn_transformer_test.py
@@ -156,7 +156,7 @@ class Test_Transformer():
                            "B": ['a', 'a', 'a', 'b']})
         df_expect = pd.DataFrame({"A": [0, 1, 2, 3],
                                   "B": [0, 0, 0, 1]})
-        ec = skex.MultiLabelEncoder()
+        ec = skex.MultiLabelEncoder(dtype=np.int64)
         df_t = ec.fit_transform(df.copy())
         assert np.where(df_expect.values == df_t.values, 0, 1).sum() == 0
         assert all(df_t.dtypes == pd.Series(dict(A=np.int64, B=np.int64)))
